+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Updated the `pdb` object, script and the generated stubs according to changes in GIMP 3.0.0-RC3.
 * Added more information to several parameter types in the generated stub file. For example, numeric parameters now contain minimum and maximum values (if different from the default minimum and maximum values), `Gio.File` indicate whether they are files or folders for opening/saving, etc.
+* Object array parameters (images, drawables, ...) are now annotated in the stub file as lists.
 * Updated type annotations for parameters that can be `None` (e.g. `Gimp.Image` or `Gimp.Layer`) in the stub file.
 * Slightly optimized access to PDB procedures and GEGL operations via the `pdb` object.
 * Improved GUI for the output directory for the stub generator.